requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
python-dotenv==1.0.1
